relocate socketio to root + recreate index .js for api app file
