+++ package.json
@@ -8,7 +8,7 @@
     "test": "jest",
     "start": "node listen.ts",
     "seed-prod": "NODE_ENV=production npm run seed",
-    "dev": "nodemon ./api/index.ts"
+    "dev": "nodemon ./index.ts"
   },
   "keywords": [],
   "author": "",
